split three objects into three files
